Fix: Use consistent local date for prices and invoices
The application was using two different methods for generating the current date, causing a bug where prices saved in one timezone would not be found by the invoice screen.

- `PriceEntryScreen` used a custom function that returned the local date.
- `CreateInvoiceScreen` used `Date.toISOString()`, which returned the UTC date.

This commit fixes the issue by creating a shared utility function, `getTodayDate`, which returns the local date in `YYYY-MM-DD` format. Both screens have been updated to use this consistent function, ensuring that prices are saved and retrieved using the same local date.
